split up main file and cleanup up old non-public endpoints
